+++ lib/persistent.js
@@ -5,16 +5,10 @@ const Cache = require('xache')
 const RecordCache = require('record-cache')
 const messages = require('./messages')
 const { SEQ_REUSED, SEQ_TOO_LOW } = require('./errors')
-const { NS_SIGNATURE } = require('./utilities')
+const { NS_ANNOUNCE, NS_UNANNOUNCE, NS_MUTABLE_PUT } = require('./ns')
 
 const TMP = Buffer.allocUnsafe(32)
 const SIGNATURE_TIMEOUT = 1 * 60 * 1000
-
-const NS_ANNOUNCE = Buffer.allocUnsafe(32)
-const NS_UNANNOUNCE = Buffer.allocUnsafe(32)
-
-sodium.crypto_generichash(NS_ANNOUNCE, Buffer.from('hyperswarm_announce'))
-sodium.crypto_generichash(NS_UNANNOUNCE, Buffer.from('hyperswarm_unannounce'))
 
 const rawArray = cenc.array(cenc.raw)
 
@@ -35,15 +29,9 @@ module.exports = class PersistentNode {
     return this.dht.id.equals(TMP)
   }
 
-  onget (req, { mutable }) {
+  onmutableget (req) {
     if (!req.target) return req.reply(null)
-    const store = mutable ? this.mutables : this.immutables
-    if (mutable === false) {
-      const k = req.target.toString('hex')
-      const value = store.get(k)
-      if (!value) return req.reply(null)
-      return req.reply(value)
-    }
+    const store = this.mutables
     const seq = cenc.decode(cenc.uint, req.value)
     const k = req.target.toString('hex')
     const value = store.get(k)
@@ -53,26 +41,18 @@ module.exports = class PersistentNode {
     return req.reply(value)
   }
 
-  onput (req, { mutable }) {
+  onmutableput (req) {
     if (!req.target || !req.token || !req.value) return // timeout
-    const store = mutable ? this.mutables : this.immutables
-    if (mutable === false) {
-      const hash = Buffer.alloc(32)
-      sodium.crypto_generichash(hash, req.value)
-      if (hash.equals(req.target) === false) return // timeout
-      const k = req.target.toString('hex')
-      store.set(k, req.value)
-      return req.reply(null)
-    }
+    const store = this.mutables
     const { seq, value, signature, publicKey } = cenc.decode(messages.mutable, req.value)
-    const hash = Buffer.alloc(32)
+    const hash = Buffer.allocUnsafe(32)
     sodium.crypto_generichash(hash, publicKey)
     if (hash.equals(req.target) === false) return // timeout
-    if (!value || !signature) return // timeout
+    if (!value || signature.length === 0) return // timeout
     const k = hash.toString('hex')
     const local = store.get(k)
     const signable = Buffer.allocUnsafe(32)
-    sodium.crypto_generichash(signable, cenc.encode(messages.signable, { value, seq }), NS_SIGNATURE)
+    sodium.crypto_generichash(signable, cenc.encode(messages.signable, { value, seq }), NS_MUTABLE_PUT)
     const verified = sodium.crypto_sign_verify_detached(signature, signable, publicKey)
     if (verified === false) return // timeout
     if (local) {
@@ -84,6 +64,26 @@ module.exports = class PersistentNode {
 
     store.set(k, req.value)
     return req.reply(req.value)
+  }
+
+  onimmutableget (req) {
+    if (!req.target) return req.reply(null)
+    const store = this.immutables
+    const k = req.target.toString('hex')
+    const value = store.get(k)
+    if (!value) return req.reply(null)
+    return req.reply(value)
+  }
+
+  onimmutableput (req) {
+    if (!req.target || !req.token || !req.value) return // timeout
+    const store = this.immutables
+    const hash = Buffer.alloc(32)
+    sodium.crypto_generichash(hash, req.value)
+    if (hash.equals(req.target) === false) return // timeout
+    const k = req.target.toString('hex')
+    store.set(k, req.value)
+    return req.reply(null)
   }
 
   onlookup (req) {
